Cleaned up code #70

+++ src/Auth/Hash.php
@@ -31,8 +31,8 @@ class Hash {
      * @since   LRS 3.28.0  Moved from Functions to `Framework\Auth\Hash`.
      */
 
-    public static function random_id_string( int $number_of_chars = 7, string $input = '' ): string {
-        return substr( md5( rand() . $input ), 0, $number_of_chars );
+    public static function random_id_string(int $number_of_chars = 7, string $input = ''): string {
+        return substr(md5(rand() . $input), 0, $number_of_chars);
     }
 
 
@@ -50,10 +50,10 @@ class Hash {
      * @since   LRS 3.28.0  Moved from Functions to `Framework\Auth\Hash`.
      */
 
-    public static function generate_guid( bool $trim = true ): string {
+    public static function generate_guid(bool $trim = true): string {
         // Windows
-        if ( function_exists( 'com_create_guid' ) === true ) {
-            if ( $trim === true ) {
+        if (function_exists('com_create_guid') === true) {
+            if ($trim === true) {
                 return trim(com_create_guid(), '{}');
             } else {
                 return com_create_guid();
@@ -61,26 +61,26 @@ class Hash {
         }
 
         // OSX / Linux
-        if ( function_exists( 'openssl_random_pseudo_bytes' ) === true ) {
-            $data = openssl_random_pseudo_bytes( 16 );
-            $data[6] = chr( ord( $data[6] ) & 0x0f | 0x40 );    // set version to 0100
-            $data[8] = chr( ord( $data[8] ) & 0x3f | 0x80 );    // set bits 6-7 to 10
-            return vsprintf( '%s%s-%s-%s-%s-%s%s%s', str_split( bin2hex( $data ), 4 ) );
+        if (function_exists('openssl_random_pseudo_bytes') === true) {
+            $data = openssl_random_pseudo_bytes(16);
+            $data[6] = chr(ord($data[6]) & 0x0f | 0x40);    // set version to 0100
+            $data[8] = chr(ord($data[8]) & 0x3f | 0x80);    // set bits 6-7 to 10
+            return vsprintf('%s%s-%s-%s-%s-%s%s%s', str_split(bin2hex($data), 4));
         }
 
         // Fallback (PHP 4.2+)
-        mt_srand( (double)microtime() * 10000 );
-        $charid = strtolower( md5( uniqid( rand(), true ) ) );
-        $hyphen = chr( 45 );                  // "-"
-        $lbrace = $trim ? "" : chr( 123 );    // "{"
-        $rbrace = $trim ? "" : chr( 125 );    // "}"
+        mt_srand((float)microtime() * 10000);
+        $charid = strtolower(md5(uniqid(rand(), true)));
+        $hyphen = chr(45);                  // "-"
+        $lbrace = $trim ? "" : chr(123);    // "{"
+        $rbrace = $trim ? "" : chr(125);    // "}"
         return $lbrace .
-                substr( $charid,  0,  8 ) . $hyphen .
-                substr( $charid,  8,  4 ) . $hyphen .
-                substr( $charid, 12,  4 ) . $hyphen .
-                substr( $charid, 16,  4 ) . $hyphen .
-                substr( $charid, 20, 12 ) . 
-                $rbrace;
+            substr($charid,  0,  8) . $hyphen .
+            substr($charid,  8,  4) . $hyphen .
+            substr($charid, 12,  4) . $hyphen .
+            substr($charid, 16,  4) . $hyphen .
+            substr($charid, 20, 12) .
+            $rbrace;
     }
 
 
@@ -97,13 +97,13 @@ class Hash {
      * @since   LRS 3.28.0
      */
 
-    public static function generate_cookie_hash( string $part1, string $part2, string $part3 ): string {
-        return md5( 
-            date( 'Y-m-d G:i:s' ) . 
-            time() . 
-            $part1 . 
-            $part2 . 
-            $part3 
+    public static function generate_cookie_hash(string $part1, string $part2, string $part3): string {
+        return md5(
+            date('Y-m-d G:i:s') .
+                time() .
+                $part1 .
+                $part2 .
+                $part3
         );
     }
 
@@ -120,13 +120,12 @@ class Hash {
      * @since   LRS 3.28.0
      */
 
-    public static function generate_session_hash( string $part1, string $part2 ): string {
-        return md5( 
-            date( 'Y-m-d G:i:s' ) . 
-            time() . 
-            $part1 . 
-            $part2 
+    public static function generate_session_hash(string $part1, string $part2): string {
+        return md5(
+            date('Y-m-d G:i:s') .
+                time() .
+                $part1 .
+                $part2
         );
     }
-
 }
